Don't let the cursor go before the initial marker (e.g. ':' or '/') when going back in the input bar.
Pressing <Ctrl-b> while the cursor is at position 1 causes it to go before the
first character (typically ':' or '/'), which is not supposed to happen, I
think.

Also, due to not binding GDK_KEY_Left as an input bar shortcut, the same thing
happens with the left arrow key, even after fixing it in
girara_isc_string_manipulation. Therefore, I bound it here (together with
GDK_KEY_Right, just for the sake of consistency).

Signed-off-by: Sebastian Ramacher <sebastian+dev@ramacher.at>