Make command history behave more like vim's
When returning to the bottom of the history, go back to what was on the
command-line.

And only match history entries starting with what has already been typed
on the command-line. For instance, if you type ":set" and the hit <Up>,
you'll go through previous ":set" commands and won't be shown any of the
previous ":map" commands.

Signed-off-by: Sebastian Ramacher <sebastian+dev@ramacher.at>