Finer-grained control for showing/hiding the scrollbars.
This patch adds two new settings: show-h-scrollbar and show-v-scrollbar, for
showing/hiding the horizontal/vertical scrollbars independently of each others.

Changing show-scrollbars affects the values of both of these settings
accordingly, and vice versa.

This is useful, for example, when reading a zoomed-in document that is centered
horizontally; if the contents of the document is visible across the width of
the view, then the horizontal scrollbar is rarely used, and by hiding it, it
gives the impression that the document is not scaled, while still allowing for
scrolling vertically using the vertical scrollbar, if desired.

Signed-off-by: Sebastian Ramacher <sebastian+dev@ramacher.at>