Do not toggle scrollbar visibility through scroll policy in GTK3
This fixes odd window size issues with gtk3. Prevously the scrollbars
where displayed or hidden by changing the scroll policy of
GtkScrolledWindow. It seems that in gtk3 a NEVER policy is understood as
no scrolling, so the container fits the size request of the page widget.

Now we set an AUTOMATIC scroll policy, and show or hide scrollbars
simply by changing the visibility of their widget.

For some reason, changing scrollbar widget visibility does not work in
GTK2, so I had handle GTK2 and GTK3 separately.

Signed-off-by: Sebastian Ramacher <sebastian+dev@ramacher.at>