make the bottom box an overlay on top of view
On GTK3, we make the bottom_box container that groups status, input
notification and completions, into an overlay on top of the main view.

This means that hiding the inputbar no loger triggers a content resize,
and we will be able, eventually, to get rid of some ugly hacks elsewhere.

Signed-off-by: Sebastian Ramacher <sebastian+dev@ramacher.at>