add config options for vertical and horizontal padding in statusbar
Adds statusbar-h-padding and statusbar-v-padding options replacing
statusbar-padding, to adjust vertical and horizontal padding for the
status, input and notification bars.

In addition, we remove a GTK2 hack introduced in commit a026b72b297b7d8
for forcing the statusbar and inputbar to have the same height. Now we
adjust the statusbar items padding directly in girara_statusbar_item_add.

Signed-off-by: Sebastian Ramacher <sebastian@ramacher.at>