Add optional support for libnotify notifications.
Added suport for libnotify as a compile-time dependency with the WITH_LIBNOTIFY
flag, enabled by default. When not enabled, any use of the function will result
in a girara style notification to the user explaining him that girara was
compiled without it.

The notification will try to use the icon in the following order:
 - If there is a "named icon" such as the ones in /usr/share/icons, it will try
   to pick it up.
 - If there is an icon file, it will try to pick it up.
 - If there is nothing, it will use the "info" stock icon.

Refer to the "window-icon" and it associated callback for more information on
icon handling.

Signed-off-by: Benoît Taine <ork@olol.eu>