use CSS syntax for default font specification
GTK3 deprecated Pango-style font specifications ("family style size") in
favor of CSS-style specifications ("style size family", for example).
However, the compatibility layer for Pango specifications is broken and
causes a parsing failure in the CSS stylesheet which girara provides.

To both fix this and make girara compatible with future GTK versions
which drop the deprecated syntax, introduce a default which uses the new
syntax.