Support Unicode characters in mappings
This patch makes it possible to use UTF-8 encoded Unicode characters in
`map` commands. Previously, it was impossible to do so because there was
an assumption that each character is always exactly one byte.

Now commands are parsed as sequences of Unicode codepoints, then the
codepoints are converted into keyvals and then treated as regular keys.